+++ core/src/main/java/com/warhammer/npc/generator/hero/repository/NicknameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NicknameRepository extends CrudRepository<Nickname, Integer> {
+public interface NicknameRepository extends CrudRepository<Nickname, Long> {
 }
